requirements/dev.txt
-r base.txt

# Development Tools
black==24.2.0
flake8==7.0.0
pytest==8.0.2
pytest-django==4.8.0
coverage==7.4.3
django-debug-toolbar==4.3.0 